Remove export vacancies to bigquery task
The task that export vacancies to big query is broken and we now use a 
different task for doing that.
