--- app/jobs/export_vacancy_records_to_big_query_job.rb
@@ -1,9 +1,0 @@
-require 'export_vacancy_records_to_big_query'
-
-class ExportVacancyRecordsToBigQueryJob < ApplicationJob
-  queue_as :export_vacancies
-
-  def perform
-    ExportVacancyRecordsToBigQuery.new.run!
-  end
-end
