+++ lib/tasks/update_database_records_in_big_query.rake
@@ -1,8 +1,7 @@
 namespace :database_records do
-  desc 'Updates Big Query Datasets with latest database records'
+  desc 'Updates Big Query User records Dataset with latest records from DSI'
   namespace :in_big_query do
     task update: :environment do
-      ExportVacancyRecordsToBigQueryJob.perform_later
       ExportUserRecordsToBigQueryJob.perform_later
     end
   end
