+++ spec/lib/tasks/update_database_records_in_big_query_rake_spec.rb
@@ -6,7 +6,6 @@ RSpec.describe 'rake database_records:in_big_query:update', type: :task do
   end
 
   it 'queues the export vacancy records to big query job' do
-    expect { task.execute }.to have_enqueued_job(ExportVacancyRecordsToBigQueryJob)
     expect { task.execute }.to have_enqueued_job(ExportUserRecordsToBigQueryJob)
   end
 end
